Move Hippo File and Hippo File parser and Rackspace CDN to lib
* This classes are use on the bin/migrate_files.rb command
